+++ CHANGELOG.md
@@ -12,12 +12,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `DatasetRemixStep` in `tango.steps`.
 - Added module `tango.common.sequences`.
 - Added `DatasetDict` class in `tango.common.dataset_dict`.
+- Added [🤗 Datasets](https://github.com/huggingface/datasets) integration.
 
 ### Changed
 
 - `tango run` command will acquire a lock on the directory to avoid race conditions.
 - Integrations can now be installed with `pip install tango[INTEGRATION_NAME]`. For example,
   `pip install tango[torch]`.
+
+### Fixed
+
+- `Step` implementations can now take arbitrary `**kwargs` in their `run()` methods.
 
 ## [v0.0.3](https://github.com/allenai/tango/releases/tag/v0.0.3) - 2021-09-27